Minor fixes [38]

+++ run_on_coal/Managers/RenderManager.h
@@ -61,7 +61,7 @@ public:
     void ClearRenderArea(GLbitfield f_params);
     static void SetClearColour(glm::vec4 &f_color) { glClearColor(f_color.r, f_color.g, f_color.b, f_color.a); }
     inline void SetViewport(glm::ivec4 &f_area) { if(!m_locked) glViewport(f_area.r, f_area.g, f_area.b, f_area.a); }
-    static void SetPolygonMode(unsigned int f_mode) { glPolygonMode(GL_FRONT_AND_BACK, GL_POINT + static_cast<int>(f_mode)); }
+    static void SetPolygonMode(int f_mode) { glPolygonMode(GL_FRONT_AND_BACK, GL_POINT + f_mode); }
 
     void SetActiveScene(Scene *f_scene);
     inline void RemoveAsActiveScene(Scene *f_scene) { if(m_activeScene == f_scene) m_activeScene = NULL; }
